Module lay::style

Styling utilities.

This module contains utilities to work with terminal CSIs and styled types.

println!(
    "{on_black}{green}{blink}{weight}# Styling utilities.{reset_blink}

    {blue}{slant}This module contains utilities to work with terminal CSIs and styled types.{reset}",
    on_black = Background(Rgb(0, 0, 0)),
    green = Foreground(Green),
    blue = Foreground(Blue),
    blink = Fast,
    reset_blink = ResetBlink,
    weight = Bold,
    slant = Italic,
    reset = Reset,
);

$ cargo run --quiet --example style01

Color

Attributes

Reset

Style

StylerIndex

Styler

Styled

Color

The Color enum is surely no surprises for you!

It lists all the available colors of the terminal, with dark variants, and with Rgb, Ansi and ResetColor variants.

As all enums in lay, its variants are re-exported:

assert_eq!(White, Color::White);

It does not Displays by itself though. Read on!

Attributes

You can use the following types to print CSIs to the terminal:

Foreground tuple struct: Foreground(Color)

Background tuple struct: Background(Color)

Weight enum: Bold, Light, ResetWeight

Slant enum: Italic, ResetSlant

Blink enum: Slow, Fast, ResetBlink

Invert enum: Inverted, ResetInvert

Strike enum: Striked, ResetStrike

Underline enum: Underlined, ResetUnderline

Overline enum: Overlined, ResetOverline

Border enum: Frame, Circle, ResetBorder

They Display the CSI they represent. Some basic examples:

println!("{}Red.{} Not red.", Foreground(Red), Foreground(ResetColor));
println!(
    "{}On Green{}. Not on green.",
    Background(Green),
    Background(ResetColor)
);
println!("{}Bold{}. Not bold.", Bold, ResetWeight);

$ cargo run --quiet --example style02

Reset

The Reset unit struct will reset all attributes sent to the screen.

Note that "reseting" (Reset and Reset* variants) means going back to the user's default attribute(s).

println!(
    "{}{}{}Multiple attributes, one reset.{} Not styled.",
    Foreground(Red),
    Background(Green),
    Bold,
    Reset
);

$ cargo run --quiet --example style03

Easy, right?

Style

Enter the Style type. A Style holds an Option for each attribute.

Why are we wrapping with Option? To convey ideas such as 'undefined' (display no CSI) or 'inherit' (inherit from some parent attribute, if any).

It also comes with a NONE (aka Default) and a RESET consts.

println!("{}Printing with style.", Style {
    foreground: Some(Foreground(White)),
    background: Some(Background(Black)),
    weight: Some(Bold),
    slant: Some(Italic),
    ..Style::NONE
});

$ cargo run --quiet --example style04

The Styler* traits make it easier to declare and use Styles. Here's a foretaste:

println!(
    "{}Declaring with style",
    Style::NONE.white().on_black().bold().italic() // Same as above
);

$ cargo run --quiet --example style05

More on that below!

StylerIndex

The StylerIndex trait defines getters for types with Optional attributes.

It will get you clones of Options of attribute from shared references:

let style = Style {
    foreground: Some(Foreground(Red)),
    weight: Some(Bold),
    ..Style::NONE
};

assert_eq!(style.get_foreground(), Some(Foreground(Red)));
assert_eq!(style.get_weight(), Some(Bold));
assert_eq!(style.get_slant(), None);

Its little brother StylerIndexMut will get you unique references to Options of attribute from unique references:

let style = &mut Style {
    foreground: Some(Foreground(Red)),
    weight: Some(Bold),
    ..Style::NONE
};

assert_eq!(style.get_foreground_mut(), &mut Some(Foreground(Red)));
assert_eq!(style.get_weight_mut(), &mut Some(Bold));
assert_eq!(style.get_slant_mut(), &mut None);

Styler

There are traits to set Optional attributes as well: Styler and StylerMut.

Even better, you can set variants or None an attribute straight as provided methods!

// Styler
println!(
    "{}Sweet!",
    Style::NONE
        .weight(Bold)  // Sets an attribute
        .no_slant()    // Nones an attribute
        .underlined()  // Sets a variant
        .reset_blink() // That's a variant as well
        .white()       // Sets a foreground color
        .on_dark_red() // Sets a background color
);

// StylerMut
let style = &mut Style::NONE;
style.weight_mut(Bold); //  Sets an attribute
style.no_slant_mut(); //    Nones an attribute
style.underlined_mut(); //  Sets a variant
style.reset_blink_mut(); // That's a variant as well
style.white_mut(); //       Sets a foreground color
style.on_dark_red_mut(); // Sets a background color

println!("{}Sick!", style);

$ cargo run --quiet --example style06

Find out some more provided methods on their own documentation!

Styler's setters return the Output associated type so that this trait is not limited to the Style type. Actually, lay's types implement the Styler* traits anywhere applicable.

Styled

We are going back to the type world to close this tour of style with Styled. Styled associates Displayable content to Styles, so it can be printed on its own.

println!("{}", Styled::from("Wow!").red().on_blue().bold());

$ cargo run --quiet --example style07

What is the point you ask? It only resets the attributes that were used:

println!(
    "{}{}, still underlined.",
    Underlined,
    Styled::from("Red").red(),
);

$ cargo run --quiet --example style08

Here is a more contrived example:

/// Makes a list
fn list(title: &str, items: &[&str], selected: usize) -> String {
    let mut list = String::from(title) + ":\n";

    for (i, item) in items.iter().enumerate() {
        if i == selected {
            // Red is flashy, right?
            list = format!("{}‣ {}\n", list, Styled::from(item).red());
        } else {
            list = format!("{}• {}\n", list, Styled::from(item));
        }
    }

    list
}

// Print list
println!(
    "{}",
    Styled::from(list("List", &vec!["Item", "Item", "Item"], 1))
        .red() // Red can be fancy too!
        .on_dark_yellow()
        .bold(),
);

// Wait, what?

$ cargo run --quiet --example style09

Oh, no! It does not work as expected... It is time to dive into layers!

Re-exports

pub use Weight::*;

pub use Slant::*;

pub use Underline::*;

pub use Strike::*;

pub use Overline::*;

pub use Invert::*;

pub use Blink::*;

pub use Border::*;

pub use Color::*;

Structs

Background
A Background Color.

Foreground
A Foreground Color.

Reset
Resets all attributes.

Style
Styles.

Styled
Displayable Styled content.

Enums

Blink
Blink (Slow, Fast, ResetBlink).

Border
Border (Circle, Frame, ResetBorder).

Color
A Color for Foreground and Background.

Invert
Invert (Inverted, ResetInvert).

Overline
Overline (Overlined, ResetOverline).

Slant
Slant (Italic, ResetSlant).

Strike
Strike (Striked, ResetStrike).

Underline
Underline (Underlined, ResetUnderline).

Weight
Weight (Bold, Light, ResetWeight).

Traits

Styler
Optional attributes setters.

StylerIndex
Optional attributes getters.

StylerIndexMut
Optional attributes getters, mutably.

StylerMut
Optional attributes setters, mutably.